+++ service/resource/cloudformation/adapter/error.go
@@ -1,0 +1,40 @@
+package adapter
+
+import "github.com/giantswarm/microerror"
+
+var tooManyResultsError = microerror.New("too many results")
+
+var wrongTypeError = microerror.New("wrong type")
+
+// IsWrongType asserts wrongTypeError.
+func IsWrongType(err error) bool {
+	return microerror.Cause(err) == wrongTypeError
+}
+
+var malformedAmazonAccountIDError = microerror.New("malformed amazon account ID")
+
+// IsMalformedAmazonAccountID asserts malformedAmazonAccountIDError.
+func IsMalformedAmazonAccountID(err error) bool {
+	return microerror.Cause(err) == malformedAmazonAccountIDError
+}
+
+var wrongAmazonAccountIDLengthError = microerror.New("wrong amazon account ID length")
+
+// IsWrongAmazonIDLength asserts wrongAmazonAccountIDLengthError.
+func IsWrongAmazonAccountIDLength(err error) bool {
+	return microerror.Cause(err) == wrongAmazonAccountIDLengthError
+}
+
+var emptyAmazonAccountIDError = microerror.New("empty amazon account ID")
+
+// IsEmptyAmazonAccountID asserts emptyAmazonAccountIDError.
+func IsEmptyAmazonAccountID(err error) bool {
+	return microerror.Cause(err) == emptyAmazonAccountIDError
+}
+
+var invalidConfigError = microerror.New("invalid config")
+
+// IsInvalidConfig asserts invalidConfigError.
+func IsInvalidConfig(err error) bool {
+	return microerror.Cause(err) == invalidConfigError
+}
