+++ service/resource/cloudformation/adapter/iam.go
@@ -1,4 +1,4 @@
-package cloudformation
+package adapter
 
 import (
 	"regexp"
