+++ service/resource/cloudformation/adapter/iam_test.go
@@ -1,4 +1,4 @@
-package cloudformation
+package adapter
 
 import (
 	"fmt"
