+++ service/resource/cloudformation/adapter/launch_configuration.go
@@ -1,27 +1,18 @@
-package cloudformation
+package adapter
 
 import (
 	"fmt"
-	"strconv"
 	"strings"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/iam"
+	"github.com/giantswarm/aws-operator/service/key"
 	"github.com/giantswarm/awstpr"
 	"github.com/giantswarm/microerror"
-
-	"github.com/giantswarm/aws-operator/service/key"
 )
 
-type hydrater func(awstpr.CustomObject, Clients) error
-
-type adapter struct {
-	ASGType string
-
-	launchConfigAdapter
-	autoScalingGroupAdapter
-}
+// template related to this adapter: service/templates/cloudformation/launch_configuration.go
 
 type launchConfigAdapter struct {
 	AssociatePublicIPAddress bool
@@ -38,37 +29,6 @@ type BlockDeviceMapping struct {
 	DeviceName          string
 	VolumeSize          int
 	VolumeType          string
-}
-
-type autoScalingGroupAdapter struct {
-	ASGMaxSize             int
-	ASGMinSize             int
-	AZ                     string
-	HealthCheckGracePeriod int
-	LoadBalancerName       string
-	MaxBatchSize           string
-	MinInstancesInService  string
-	RollingUpdatePauseTime string
-	SubnetID               string
-}
-
-func newAdapter(customObject awstpr.CustomObject, clients Clients) (adapter, error) {
-	a := adapter{}
-
-	a.ASGType = prefixWorker
-
-	hydraters := []hydrater{
-		a.getAutoScalingGroup,
-		a.getLaunchConfiguration,
-	}
-
-	for _, h := range hydraters {
-		if err := h(customObject, clients); err != nil {
-			return adapter{}, microerror.Mask(err)
-		}
-	}
-
-	return a, nil
 }
 
 func (l *launchConfigAdapter) getLaunchConfiguration(customObject awstpr.CustomObject, clients Clients) error {
@@ -143,52 +103,6 @@ func (l *launchConfigAdapter) getLaunchConfiguration(customObject awstpr.CustomO
 		return microerror.Mask(err)
 	}
 	l.SmallCloudConfig = smallCloudConfig
-
-	return nil
-}
-
-func (a *autoScalingGroupAdapter) getAutoScalingGroup(customObject awstpr.CustomObject, clients Clients) error {
-	a.AZ = customObject.Spec.AWS.AZ
-	workers := key.WorkerCount(customObject)
-	a.ASGMaxSize = workers
-	a.ASGMinSize = workers
-	a.MaxBatchSize = strconv.FormatFloat(asgMaxBatchSizeRatio, 'f', -1, 32)
-	a.MinInstancesInService = strconv.FormatFloat(asgMinInstancesRatio, 'f', -1, 32)
-	a.HealthCheckGracePeriod = gracePeriodSeconds
-	a.RollingUpdatePauseTime = rollingUpdatePauseTime
-
-	// load balancer name
-	// TODO: remove this code once the ingress load balancer is created by cloudformation
-	// and add a reference in the template
-	lbName, err := key.LoadBalancerName(customObject.Spec.Cluster.Kubernetes.IngressController.Domain, customObject)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-	a.LoadBalancerName = lbName
-
-	// subnet ID
-	// TODO: remove this code once the subnet is created by cloudformation and add a
-	// reference in the template
-	subnetName := key.SubnetName(customObject, suffixPublic)
-	describeSubnetInput := &ec2.DescribeSubnetsInput{
-		Filters: []*ec2.Filter{
-			{
-				Name: aws.String(fmt.Sprintf("tag:%s", tagKeyName)),
-				Values: []*string{
-					aws.String(subnetName),
-				},
-			},
-		},
-	}
-	output, err := clients.EC2.DescribeSubnets(describeSubnetInput)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-	if len(output.Subnets) > 1 {
-		return microerror.Mask(tooManyResultsError)
-	}
-
-	a.SubnetID = *output.Subnets[0].SubnetId
 
 	return nil
 }
