+++ service/resource/cloudformation/error.go
@@ -32,32 +32,9 @@ func IsStackNotFound(err error) bool {
 	return strings.Contains(microerror.Cause(err).Error(), "does not exist")
 }
 
-var tooManyResultsError = microerror.New("too many results")
-
 var wrongTypeError = microerror.New("wrong type")
 
 // IsWrongType asserts wrongTypeError.
 func IsWrongType(err error) bool {
 	return microerror.Cause(err) == wrongTypeError
-}
-
-var malformedAmazonAccountIDError = microerror.New("malformed amazon account ID")
-
-// IsMalformedAmazonAccountID asserts malformedAmazonAccountIDError.
-func IsMalformedAmazonAccountID(err error) bool {
-	return microerror.Cause(err) == malformedAmazonAccountIDError
-}
-
-var wrongAmazonAccountIDLengthError = microerror.New("wrong amazon account ID length")
-
-// IsWrongAmazonIDLength asserts wrongAmazonAccountIDLengthError.
-func IsWrongAmazonAccountIDLength(err error) bool {
-	return microerror.Cause(err) == wrongAmazonAccountIDLengthError
-}
-
-var emptyAmazonAccountIDError = microerror.New("empty amazon account ID")
-
-// IsEmptyAmazonAccountID asserts emptyAmazonAccountIDError.
-func IsEmptyAmazonAccountID(err error) bool {
-	return microerror.Cause(err) == emptyAmazonAccountIDError
 }
