+++ service/resource/legacy/machine.go
@@ -14,7 +14,7 @@ import (
 	awsutil "github.com/giantswarm/aws-operator/client/aws"
 	"github.com/giantswarm/aws-operator/resources"
 	awsresources "github.com/giantswarm/aws-operator/resources/aws"
-	"github.com/giantswarm/aws-operator/service/resource/cloudformation"
+	"github.com/giantswarm/aws-operator/service/resource/cloudformation/adapter"
 )
 
 type instanceNameInput struct {
@@ -168,7 +168,7 @@ func (s *Resource) runMachine(input runMachineInput) (bool, string, error) {
 	// cloudconfig" that just fetches the previously uploaded "final
 	// cloudconfig" and executes coreos-cloudinit with it as argument.
 	// We do this to circumvent the 16KB limit on user-data for EC2 instances.
-	cloudconfigConfig := cloudformation.SmallCloudconfigConfig{
+	cloudconfigConfig := adapter.SmallCloudconfigConfig{
 		MachineType: input.prefix,
 		Region:      input.cluster.Spec.AWS.Region,
 		S3URI:       s.bucketName(input.cluster),
@@ -185,7 +185,7 @@ func (s *Resource) runMachine(input runMachineInput) (bool, string, error) {
 		return false, "", microerror.Mask(err)
 	}
 
-	smallCloudconfig, err := cloudformation.SmallCloudconfig(cloudconfigConfig)
+	smallCloudconfig, err := adapter.SmallCloudconfig(cloudconfigConfig)
 	if err != nil {
 		return false, "", microerror.Mask(err)
 	}
